requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0